refactored ExampleView
